Refs #268. Implemented all the BltSessionXxx functions.
git-svn-id: https://svn.code.sf.net/p/openblt/code/branches/hostlib@245 5dc33758-31d5-4daf-9ae8-b24bf3d40d73

+++ Host/Source/LibOpenBLT/openblt.h
@@ -110,11 +110,11 @@ LIBOPENBLT_EXPORT char const * BltVersionGetString(void);
 /** \brief Structure layout of the XCP version 1.0 session settings. */
 typedef struct t_blt_session_settings_xcp_v10
 {
-  uint32_t timeoutT1;            /**< Command response timeout in milliseconds.        */
-  uint32_t timeoutT3;            /**< Start programming timeout in milliseconds.       */
-  uint32_t timeoutT4;            /**< Erase memory timeout in milliseonds.             */
-  uint32_t timeoutT5;            /**< Program memory and reset timeout in milliseonds. */
-  uint32_t timeoutT7;            /**< Busy wait timer timeout in milliseonds.          */
+  uint16_t timeoutT1;            /**< Command response timeout in milliseconds.        */
+  uint16_t timeoutT3;            /**< Start programming timeout in milliseconds.       */
+  uint16_t timeoutT4;            /**< Erase memory timeout in milliseonds.             */
+  uint16_t timeoutT5;            /**< Program memory and reset timeout in milliseonds. */
+  uint16_t timeoutT7;            /**< Busy wait timer timeout in milliseonds.          */
   char const * seedKeyFile;      /**< Seed/key algorithm library filename.             */
 } tBltSessionSettingsXcpV10;
